requirements.txt
pytest==6.1.2
requests
allure-pytest